news-mosaic/backend/requirements.txt
fastapi
uvicorn[standard]
pydantic>=2
python-dotenv
httpx
numpy
scikit-learn
google-generativeai
nltk
